refactor: cleanup `NodeActionsMenu`
