+++ packages/shared/components/modals/index.js
@@ -2,7 +2,6 @@ export { default as AssetActionsMenu } from './AssetActionsMenu.svelte'
 export { default as CollectibleDetailsMenu } from './CollectibleDetailsMenu.svelte'
 export { default as ExpirationTimePickerMenu } from './ExpirationTimePickerMenu.svelte'
 export { default as Modal } from './Modal.svelte'
-export { default as NodeActionsMenu } from './NodeActionsMenu.svelte'
 export { default as ProfileActionsModal } from './ProfileActions.svelte'
 export { default as ProposalsDetailsMenu } from './ProposalsDetailsMenu.svelte'
 export { default as RecipientAccountSelector } from './RecipientAccountSelector.svelte'
